Update keep.R
Updated keep function in order to work inside functions, thus removing and keeping variables of the function environment in which it is called. For example:

foo <- function() {
  
  a <- 1
  b <- 2
  c <- 3
  keep(a)
  print(a)
  print(b)
  
}

Calling foo() will generate an error since "b" is deleted, only "a" is kept inside foo().

This also works for the global env:
a <- 1
b <- 2
c <- 3
keep(a)
print(a)
print(b)
ls()

will generate an error when trying to print the non-existing "b", and list only "a" in the environment.